Clean up displays folder
