+++ src/components/display/ClimateClock.tsx
