+++ src/components/display/ImageWithCaption.tsx
